hopefully fix more invalidations (#139)
* hopefully fix more invalidations

* more assertions on readbytes!

+++ src/extract.jl
@@ -576,7 +576,7 @@ function read_standard_header(
     # zero block indicates end of tarball
     if all(iszero, data)
         while !eof(io)
-            r = readbytes!(io, buf)
+            r = readbytes!(io, buf)::Integer
             write(tee, view(buf, 1:r))
         end
         return nothing
@@ -701,7 +701,7 @@ function read_data(
     padded_size = round_up(size)
     while padded_size > 0
         max_read_len = Int(min(padded_size, length(buf)))
-        read_len = readbytes!(tar, buf, max_read_len)
+        read_len = readbytes!(tar, buf, max_read_len)::Integer
         write(tee, view(buf, 1:read_len))
         read_len < max_read_len && eof(tar) && throw(EOFError())
         size -= write(file, view(buf, 1:Int(min(read_len, size))))
